Refactor tsv helper functions to new tsv module
